requirements.txt
bcrypt
colorama
pycryptodome
argparse
